+++ CHANGELOG.md
@@ -106,6 +106,10 @@
 - sctbench/inspect_examples/boundedBuffer.c: exclude `malloc.h`
 - sctbench/inspect_examples/thread-pool.example.c: return `NULL` in `tp_work_thread()`
 - sctbench/inspect_examples/ctrace-test.c: add a `void*` parameter to `trc_start_server()`
+- sctbench/inspect_examples/ctrace.foobar.comb.c: include `unistd.h`
+- sctbench/inspect_examples/ctrace.foobar.comb.c: include `stdlib.h`
+- sctbench/inspect_examples/ctrace.foobar.comb.c: declare `trc_start_client()`
+- sctbench/inspect_examples/ctrace.foobar.comb.c: add a `void*` parameter to `trc_start_server()`
 
 ## Added
 - build action for macOS